requirements.txt
numpy
scikit-learn
tensorflow==2.5.0
tensorflow-privacy==0.1.0
umap-learn